Modified TestMap_Pablo & added item
Created and added a child BP of BP_Item, BP_TempleItem.

Replaced the static mesh sphere that represented the item with BP_TempleItem.

BP_TempleItem is collectible using the south button on a controller or the E button on the keyboard. It prints a short TODO list when picked up.