Fixed bugs & added player feedback
The notes on the wall now vibrate with the player's voice when sung.

Notes that are gray are not discovered, the player needs to sing to find them. When found, they vibrate and gain their color back.

When the player finishes the puzzle an event gets dispatched to trigger other stuff.

TODO: add clues for players that have a harder time in the form of a sound bite that represents the note they have to sing